+++ src/main/java/com/athenhub/stockservice/stock/infrastructure/client/product/ProductClient.java
@@ -1,6 +1,6 @@
 package com.athenhub.stockservice.stock.infrastructure.client.product;
 
-import com.athenhub.stockservice.global.infrastructure.feignclient.FeignClientConfig;
+import com.athenhub.stockservice.stock.infrastructure.client.config.FeignClientConfig;
 import com.athenhub.stockservice.stock.infrastructure.client.product.config.ProductFeignConfig;
 import com.athenhub.stockservice.stock.infrastructure.client.product.dto.ProductDetail;
 import java.util.UUID;
