+++ src/main/java/com/athenhub/stockservice/stock/infrastructure/rabbitmq/RabbitStockProperties.java
@@ -1,4 +1,4 @@
-package com.athenhub.stockservice.global.infrastructure.rabbitmq;
+package com.athenhub.stockservice.stock.infrastructure.rabbitmq;
 
 import lombok.Data;
 import org.springframework.boot.context.properties.ConfigurationProperties;
